requirements.txt
azure_storage==0.37.0
pandas==1.5.2
